sizeof long and sizeof int might be same
memory_managers/freelists.cc: In member function 'virtual
MEDDLY::memory_manager* MEDDLY::freelist_style::initManager(unsigned
char, unsigned char, MEDDLY::memstats&) const':
memory_managers/freelists.cc:335:5: error: duplicate case value
     case sizeof(long):
     ^~~~
memory_managers/freelists.cc:332:5: note: previously used here
     case sizeof(int):
     ^~~~
memory_managers/freelists.cc: In instantiation of 'MEDDLY::node_address
MEDDLY::freelist_manager<INT>::requestChunk(size_t&) [with INT = long
int; MEDDLY::node_address = long long unsigned int; size_t = long long
unsigned int]':
memory_managers/freelists.cc:137:22:   required from here

+++ src/memory_managers/freelists.cc
@@ -332,8 +332,8 @@ MEDDLY::freelist_style::initManager(unsigned char granularity,
     case sizeof(int):
       return new freelist_manager<int>(getName(), stats);
 
-    case sizeof(long):
-      return new freelist_manager<long>(getName(), stats);
+    case sizeof(size_t):
+      return new freelist_manager<size_t>(getName(), stats);
 
     default:
       return 0;
